introduce fdroid flavor
