+++ app/src/fdroid/AndroidManifest.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest package="com.fastaccess"
+          xmlns:android="http://schemas.android.com/apk/res/android"
+          xmlns:tools="http://schemas.android.com/tools">
+    <uses-permission android:name="com.android.vending.BILLING" tools:node="remove"/>
+    <application android:name="com.fastaccess.App">
+        <activity
+            android:name="io.octo.bear.pago.BillingActivity"
+            tools:node="remove" />
+    </application>
+</manifest>
