+++ app/src/fdroid/java/com/fastaccess/ui/base/AdActivity.java
@@ -1,0 +1,27 @@
+package com.fastaccess.ui.base;
+
+import android.support.annotation.Nullable;
+import android.view.View;
+import android.widget.TextView;
+
+import com.fastaccess.R;
+import com.fastaccess.helper.PrefGetter;
+
+import net.grandcentrix.thirtyinch.TiActivity;
+import net.grandcentrix.thirtyinch.TiPresenter;
+import net.grandcentrix.thirtyinch.TiView;
+
+import butterknife.BindView;
+
+/**
+ * Created by thermatk on 12.04.17.
+ */
+
+public abstract class AdActivity<V extends TiView,P extends TiPresenter<V>> extends TiActivity<P, V> {
+
+    @Nullable @BindView(R.id.adView) TextView adView;
+
+    protected void showHideAds() {
+
+    }
+}
