+++ app/src/fdroid/res/values/secrets.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="donation_product_1">your_product_1</string>
+    <string name="donation_product_2">your_product_2</string>
+    <string name="donation_product_3">your_product_3</string>
+    <string name="donation_product_4">your_product_4</string>
+</resources>
