+++ app/src/gms/res/layout/add_banner_layout.xml
